+++ giraph-examples/src/main/java/org/apache/giraph/examples/hyperball/HyperBallTextInputFormat.java
@@ -23,7 +23,7 @@ import com.google.common.collect.Lists;
 import org.apache.giraph.edge.Edge;
 import org.apache.giraph.edge.EdgeFactory;
 import org.apache.giraph.io.formats.TextVertexInputFormat;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.NullWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.InputSplit;
@@ -34,7 +34,7 @@ import java.util.List;
 import java.util.regex.Pattern;
 
 public class HyperBallTextInputFormat extends
-  TextVertexInputFormat<IntWritable, EstimatedNF, NullWritable> {
+  TextVertexInputFormat<LongWritable, EstimatedNF, NullWritable> {
 
   /** Separator of the vertex and neighbors */
   private static final Pattern SEPARATOR = Pattern.compile("[\t ]");
@@ -50,17 +50,17 @@ public class HyperBallTextInputFormat extends
       TextVertexReaderFromEachLineProcessed<String[]> {
 
     /** Cached vertex id for the current line */
-    private IntWritable id = new IntWritable();
+    private LongWritable id = new LongWritable();
 
     @Override
     protected String[] preprocessLine(Text line) throws IOException {
       String[] tokens = SEPARATOR.split(line.toString());
-      id.set(Integer.parseInt(tokens[0]));
+      id.set(Long.parseLong(tokens[0]));
       return tokens;
     }
 
     @Override
-    protected IntWritable getId(String[] tokens) throws IOException {
+    protected LongWritable getId(String[] tokens) throws IOException {
       return id;
     }
 
@@ -70,13 +70,13 @@ public class HyperBallTextInputFormat extends
     }
 
     @Override
-    protected Iterable<Edge<IntWritable, NullWritable>> getEdges(
+    protected Iterable<Edge<LongWritable, NullWritable>> getEdges(
         String[] tokens) throws IOException {
-      List<Edge<IntWritable, NullWritable>> edges =
+      List<Edge<LongWritable, NullWritable>> edges =
           Lists.newArrayListWithCapacity(tokens.length - 1);
       for (int n = 1; n < tokens.length; n++) {
         edges.add(EdgeFactory.create(
-            new IntWritable(Integer.parseInt(tokens[n]))));
+            new LongWritable(Long.parseLong(tokens[n]))));
       }
       return edges;
     }
